refactor(commands): cleanup commands and update rule table call
